requirements.txt
Flask==2.0.1
Jinja2==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Werkzeug==2.0.1
Flask-Login==0.5.0
Flask-Bcrypt==0.7.1
python-dotenv==0.19.0
pytest==6.2.4
Flask-Testing==0.8.1
Flask-RESTful==0.3.9
Flask-WTF==0.15.1
Flask-Mail==0.9.1
Flask-CORS==3.0.10
SQLAlchemy==1.4.46  # Downgraded to avoid issues with SQLAlchemy 2.x
